Improve test
