+++ .isort.cfg
@@ -1,2 +1,2 @@
 [settings]
-known_third_party = docutils,geoh5py,numpy,properties,pytest,six,sphinx
+known_third_party = docutils,geoh5py,numpy,properties,six,sphinx
